#1 Updates made so that actions are done through API

+++ Almostengr.PetFeeder.Api/Models/Schedule.cs
@@ -12,7 +12,8 @@ namespace Almostengr.PetFeeder.Api.Models
         [Required]
         public DateTime ScheduledTime { get; set; }
         public bool IsActive { get; set; } = true;
-        public double FeedingAmount { get; set; }
+        public int FeedingAmount { get; set; }
+        // public double FeedingAmount { get; set; }
         public DayFrequency Frequency { get; set; }
         public string IpAddress { get; set; }
     }
